+++ cmd/extract.go
@@ -37,7 +37,7 @@ func mkFolder(foldername string) string {
 }
 
 // Extract extract layers from inspector tar
-func Extract(ctx context.Context, tarPath string, tarIO io.ReadCloser) (*layer.Manifest, error) {
+func Extract(ctx context.Context, tarPath string, tarIO []io.ReadCloser) (*layer.Manifest, error) {
 	var tarReader *tar.Reader
 
 	tempPath := mkFolder(RandomString())
@@ -52,17 +52,27 @@ func Extract(ctx context.Context, tarPath string, tarIO io.ReadCloser) (*layer.M
 
 		tarReader = tar.NewReader(image)
 	} else {
-		tarReader = tar.NewReader(tarIO)
-	}
-
-	err := pkg.Decompress(tarReader, tempPath)
-	if err != nil {
-		log.Printf("extract tar file failed: %v", err)
+		tarReader = tar.NewReader(tarIO[0])
 	}
 
 	// command `docker export` will generate a single file system
 	// just return the directory
 	if ctx.Value("tarType") == "container" {
+		err := pkg.Decompress(tarReader, tempPath)
+		if err != nil {
+			log.Printf("extract tar file failed: %v", err)
+		}
+
+		// Get mount path
+		for _, mio := range tarIO[1:] {
+			tarReader = tar.NewReader(mio)
+			err = pkg.Walk(tarReader, tempPath)
+			if err != nil {
+				log.Printf("decompress mount path failed, error: %v", err)
+				continue
+			}
+		}
+
 		img := &layer.Manifest{
 			Localpath: tempPath,
 			Hash:      "container",
@@ -71,32 +81,20 @@ func Extract(ctx context.Context, tarPath string, tarIO io.ReadCloser) (*layer.M
 		return img, nil
 	}
 
-	// remove all file
-	defer func() {
-		err := os.RemoveAll(tempPath)
-		if err != nil {
-			log.Printf("failed to remove %s : %v", tempPath, err)
-		}
-	}()
-
 	// need temp folder path to get layer.tar
-	img, err := Inspect(ctx, tempPath)
+	img, err := Inspect(ctx, tempPath, tarReader)
 	if err != nil {
 		log.Printf("Getting layers failed")
 		return nil, err
 	}
 
-	imgPath := mkFolder(img.Hash)
-
 	// integrate all layers
 	for _, l := range img.Layers {
-		err := l.Integration(imgPath)
+		err := l.Integration(tempPath, l.Hash)
 		if err != nil {
-			//log.Printf("layer %s integrating failed, error: %v", l.Hash, err)
 			continue
 		}
 	}
-	//log.Printf("Integrating layer successful")
 
 	return img, nil
 }