+++ pkg/inspector/utils.go
@@ -8,8 +8,9 @@ import (
 	"github.com/docker/docker/client"
 )
 
-func GetTarFromID(ctx context.Context, ID string) (io.ReadCloser, error) {
+func GetTarFromID(ctx context.Context, ID string) ([]io.ReadCloser, error) {
 	var err error
+
 	// Use the inspector id from containerd or crio
 	cli, err := client.NewClientWithOpts(client.FromEnv, client.WithAPIVersionNegotiation())
 	if err != nil {
@@ -22,7 +23,7 @@ func GetTarFromID(ctx context.Context, ID string) (io.ReadCloser, error) {
 
 	defer c.DCli.Close()
 
-	var tarFile io.ReadCloser
+	var tarFile []io.ReadCloser
 
 	if ctx.Value("tarType") == "image" {
 		tarFile, err = c.GetImageName(ID)